moving cd extraction functionality to separate file
