+++ CMakeLists.txt
@@ -75,6 +75,7 @@ add_executable(roller
     PROJECTS/ROLLER/polytex.c
     PROJECTS/ROLLER/replay.c
     PROJECTS/ROLLER/roller.c
+    PROJECTS/ROLLER/rollercd.c
     PROJECTS/ROLLER/rollercomms.c
     PROJECTS/ROLLER/sound.c
     PROJECTS/ROLLER/svgacpy.c
