+++ PROJECTS/ROLLER/rollercd.h
@@ -1,0 +1,8 @@
+#ifndef _ROLLER_ROLLERCD_H
+#define _ROLLER_ROLLERCD_H
+//-------------------------------------------------------------------------------------------------
+
+void ExtractFATDATA(const char *szImagePath, const char *szOutDir);
+
+//-------------------------------------------------------------------------------------------------
+#endif
