+++ build.zig
@@ -15,7 +15,7 @@ pub fn build(b: *std.Build) void {
     const c_flags: []const []const u8 = if (crash_debug)
         &.{ "-fwrapv", "-fno-omit-frame-pointer" }
     else
-        &.{ "-fwrapv" };
+        &.{"-fwrapv"};
 
     const exe_mod = b.createModule(.{
         .target = target,
@@ -70,6 +70,7 @@ pub fn build(b: *std.Build) void {
             "PROJECTS/ROLLER/polytex.c",
             "PROJECTS/ROLLER/replay.c",
             "PROJECTS/ROLLER/roller.c",
+            "PROJECTS/ROLLER/rollercd.c",
             "PROJECTS/ROLLER/rollercomms.c",
             "PROJECTS/ROLLER/sound.c",
             "PROJECTS/ROLLER/svgacpy.c",
